// frontend/src/api/model/coverageStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type CoverageStatus = typeof CoverageStatus[keyof typeof CoverageStatus];


export const CoverageStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  EXPIRED: 'EXPIRED',
} as const;